rename: Main 페이지 폴더 생성
